Abstract RotaryPosMixin to reuse
