+++ python/sglang/srt/models/qwen3_vl.py
@@ -50,7 +50,7 @@ from sglang.srt.managers.schedule_batch import (
 from sglang.srt.model_executor.forward_batch_info import ForwardBatch, PPProxyTensors
 from sglang.srt.model_loader.weight_utils import default_weight_loader
 from sglang.srt.models.qwen3 import Qwen3Model
-from sglang.srt.models.utils import compute_cu_seqlens_from_grid_numpy
+from sglang.srt.models.utils import RotaryPosMixin, compute_cu_seqlens_from_grid_numpy
 from sglang.srt.multimodal.mm_utils import run_dp_sharded_mrope_vision_model
 from sglang.srt.server_args import get_global_server_args
 from sglang.srt.utils import add_prefix
@@ -257,7 +257,7 @@ class Qwen3VLMoeVisionPatchMerger(nn.Module):
         return out
 
 
-class Qwen3VLMoeVisionModel(nn.Module):
+class Qwen3VLMoeVisionModel(nn.Module, RotaryPosMixin):
 
     def __init__(
         self,
@@ -336,42 +336,12 @@ class Qwen3VLMoeVisionModel(nn.Module):
     def device(self) -> torch.device:
         return self.patch_embed.proj.weight.device
 
-    @staticmethod
-    @lru_cache(maxsize=1024)
-    def rot_pos_ids(h: int, w: int, spatial_merge_size: int) -> torch.Tensor:
-        hpos_ids = np.broadcast_to(np.arange(h).reshape(h, 1), (h, w))
-        h_div = h // spatial_merge_size
-        w_div = w // spatial_merge_size
-        hpos_ids = hpos_ids.reshape(
-            h_div,
-            spatial_merge_size,
-            w_div,
-            spatial_merge_size,
-        )
-        hpos_ids = hpos_ids.transpose(0, 2, 1, 3)
-        hpos_ids = hpos_ids.flatten()
-
-        wpos_ids = np.broadcast_to(np.arange(w).reshape(1, w), (h, w))
-        wpos_ids = wpos_ids.reshape(
-            h_div,
-            spatial_merge_size,
-            w_div,
-            spatial_merge_size,
-        )
-        wpos_ids = wpos_ids.transpose(0, 2, 1, 3)
-        wpos_ids = wpos_ids.flatten()
-
-        return torch.from_numpy(np.stack([hpos_ids, wpos_ids], axis=-1))
-
     def rot_pos_emb(self, grid_thw):
-        pos_ids = [
-            (
-                self.rot_pos_ids(h, w, self.spatial_merge_size)
-                if t == 1
-                else self.rot_pos_ids(h, w, self.spatial_merge_size).repeat(t, 1)
-            )
-            for t, h, w in grid_thw
-        ]
+        pos_ids = []
+        for t, h, w in grid_thw:
+            base = self.rot_pos_ids(h, w, self.spatial_merge_size)
+            pos_ids.append(base if t == 1 else base.repeat(t, 1))
+
         pos_ids = torch.cat(pos_ids, dim=0)
         max_grid_size = grid_thw[:, 1:].max()
         rotary_pos_emb_full = self.rotary_pos_emb(max_grid_size)
